pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "maks-tabu", email = "maxtabolin@gmail.com"},
]
dependencies = [
    "polars>=0.19.13",
    "Pillow>=10.1.0",
    "transformers>=4.35.1",
    "black>=23.11.0",
    "tslearn>=0.6.2",
    "numpy>=1.25.2",
    "scikit-learn>=1.3.2",
    "datasets>=2.14.6",
    "torch==1.12.1",
    "jiwer>=3.0.3",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

